+++ lib/lightning_web/live/project_live/form_component.ex
@@ -18,26 +18,40 @@ defmodule LightningWeb.ProjectLive.FormComponent do
 
   import LightningWeb.Components.Form
 
-  import Ecto.Changeset, only: [fetch_field!: 2, put_assoc: 3]
+  import Ecto.Changeset, only: [fetch_field!: 2]
 
   @impl true
   def update(
         %{project: project, users: users} = assigns,
         socket
       ) do
-    changeset = Projects.change_project(project)
+    project_users_ids = Enum.map(project.project_users, & &1.user_id)
 
-    all_users = users |> Enum.map(&{"#{&1.first_name} #{&1.last_name}", &1.id})
+    users_without_access =
+      Enum.reject(users, fn user -> user.id in project_users_ids end)
+
+    p_users_without_access =
+      Enum.map(users_without_access, fn user ->
+        %Lightning.Projects.ProjectUser{user_id: user.id, user: user, role: nil}
+      end)
+
+    all_project_users = project.project_users ++ p_users_without_access
+
+    project_users =
+      Enum.sort_by(
+        all_project_users,
+        fn p_user -> p_user.user.first_name end,
+        :asc
+      )
+
+    changeset =
+      Projects.change_project(%{project | project_users: project_users})
 
     {:ok,
      socket
-     |> assign(assigns |> Map.drop([:users]))
+     |> assign(assigns)
      |> assign(:changeset, changeset)
-     |> assign(
-       all_users: all_users,
-       available_users: filter_available_users(changeset, all_users),
-       selected_member: ""
-     )
+     |> assign(:project_users, project_users)
      |> assign(
        :name,
        Projects.url_safe_project_name(fetch_field!(changeset, :name))
@@ -45,9 +59,16 @@ defmodule LightningWeb.ProjectLive.FormComponent do
   end
 
   @impl true
-  def handle_event("validate", %{"project" => project_params}, socket) do
+  def handle_event(
+        "validate",
+        %{"project" => project_params},
+        %{assigns: assigns} = socket
+      ) do
+    # we update the project here so that we can mantain the users in the changeset after validation
+    project = %{assigns.project | project_users: assigns.project_users}
+
     changeset =
-      socket.assigns.project
+      project
       |> Projects.change_project(
         project_params
         |> coerce_raw_name_to_safe_name()
@@ -60,94 +81,26 @@ defmodule LightningWeb.ProjectLive.FormComponent do
      |> assign(:name, fetch_field!(changeset, :name))}
   end
 
-  @impl true
-  def handle_event("delete_member", %{"index" => index}, socket) do
-    index = String.to_integer(index)
-
-    project_users_params =
-      fetch_field!(socket.assigns.changeset, :project_users)
-      |> Enum.with_index()
-      |> Enum.reduce([], fn {pu, i}, project_users ->
-        if i == index do
-          if is_nil(pu.id) do
-            project_users
-          else
-            [Ecto.Changeset.change(pu, %{delete: true}) | project_users]
-          end
-        else
-          [pu | project_users]
-        end
+  def handle_event("save", %{"project" => project_params}, socket) do
+    # Drop non-persited project users without role
+    users =
+      Enum.reject(project_params["project_users"] || %{}, fn {_key, params} ->
+        is_nil(params["id"]) and params["role"] == ""
       end)
 
-    changeset =
-      socket.assigns.changeset
-      |> put_assoc(:project_users, project_users_params)
-      |> Map.put(:action, :validate)
+    users_params =
+      Enum.map(users, fn {index, params} ->
+        if params["role"] == "" do
+          {index, Map.merge(params, %{"delete" => "true"})}
+        else
+          {index, params}
+        end
+      end)
+      |> Enum.into(%{})
 
-    available_users = filter_available_users(changeset, socket.assigns.all_users)
+    params = Map.merge(project_params, %{"project_users" => users_params})
 
-    {:noreply,
-     socket |> assign(changeset: changeset, available_users: available_users)}
-  end
-
-  @impl true
-  def handle_event(
-        "select_item",
-        %{"id" => user_id},
-        socket
-      ) do
-    {:noreply, socket |> assign(selected_member: user_id)}
-  end
-
-  @impl true
-  def handle_event(
-        "add_new_member",
-        %{"userid" => user_id},
-        socket
-      ) do
-    project_users = fetch_field!(socket.assigns.changeset, :project_users)
-
-    project_users =
-      Enum.find(project_users, fn pu -> pu.user_id == user_id end)
-      |> if do
-        project_users
-        |> Enum.map(fn pu ->
-          if pu.user_id == user_id do
-            Ecto.Changeset.change(pu, %{delete: false})
-          end
-        end)
-      else
-        project_users
-        |> Enum.concat([%Lightning.Projects.ProjectUser{user_id: user_id}])
-      end
-
-    changeset =
-      socket.assigns.changeset
-      |> put_assoc(:project_users, project_users)
-      |> Map.put(:action, :validate)
-
-    available_users = filter_available_users(changeset, socket.assigns.all_users)
-
-    {:noreply,
-     socket
-     |> assign(
-       changeset: changeset,
-       available_users: available_users,
-       selected_member: ""
-     )}
-  end
-
-  def handle_event("save", %{"project" => project_params}, socket) do
-    save_project(socket, socket.assigns.action, project_params)
-  end
-
-  defp filter_available_users(changeset, all_users) do
-    existing_ids =
-      fetch_field!(changeset, :project_users)
-      |> Enum.reject(fn pu -> pu.delete end)
-      |> Enum.map(fn pu -> pu.user_id end)
-
-    all_users |> Enum.reject(fn {_, user_id} -> user_id in existing_ids end)
+    save_project(socket, socket.assigns.action, params)
   end
 
   defp save_project(socket, :edit, project_params) do
@@ -192,13 +145,6 @@ defmodule LightningWeb.ProjectLive.FormComponent do
     end
   end
 
-  defp user_name_for_id(users, user_id) do
-    users
-    |> Enum.find_value(fn {name, id} ->
-      if id == user_id, do: name
-    end)
-  end
-
   # TODO: Determine the list of users to notify into the Project context
   # by using the changeset to determine what records are going to be added/removed
   defp filter_users_to_notify(project, project_users_params) do
@@ -235,5 +181,9 @@ defmodule LightningWeb.ProjectLive.FormComponent do
 
   defp coerce_raw_name_to_safe_name(%{} = params) do
     params
+  end
+
+  defp full_user_name(user) do
+    "#{user.first_name} #{user.last_name}"
   end
 end