+++ test/lightning/digest_email_worker_test.exs
@@ -11,10 +11,10 @@ defmodule Lightning.DigestEmailWorkerTest do
 
   describe "perform/1" do
     test "projects that are scheduled for deletion are not part of the projects for which digest alerts are sent" do
-      user = AccountsFixtures.user_fixture()
+      user = insert(:user)
 
       project =
-        ProjectsFixtures.project_fixture(
+        insert(:project,
           scheduled_deletion: Timex.now(),
           project_users: [
             %{user_id: user.id, digest: :daily}